LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-never-http-http-expected.txt

Test that 'no-referrer' works correctly from HTTP to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is ""
PASS successfullyParsed is true

TEST COMPLETE
